+++ src/main/apis/authors/author.delete.api.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { ValidError } from "../../../../features/shared/error/valid/index.js";
+import { zodValidator } from "../../../../features/shared/validation/index.js";
 import type { Context } from "../../context.js";
 import { TOKENS } from "../../di/token.js";
 
@@ -8,16 +8,7 @@ export const deleteAuthorSchema = z.object({
 });
 export type DeleteAuthorRequest = z.infer<typeof deleteAuthorSchema>;
 
-export function deleteAuthorValidator(args: unknown, ctx: Context) {
-	const logger = ctx.container.get(TOKENS.LOGGER);
-	const valid = deleteAuthorSchema.safeParse(args);
-	if (!valid.success) {
-		const error = new ValidError(valid.error);
-		logger.debug("invalid request", { error });
-		throw error;
-	}
-	return valid.data;
-}
+export const deleteAuthorValidator = zodValidator(deleteAuthorSchema);
 
 export interface DeleteAuthorResponse {
 	success: boolean;
