+++ src/main/apis/authors/author.detail.api.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { ValidError } from "../../../../features/shared/error/valid/index.js";
+import { zodValidator } from "../../../../features/shared/validation/index.js";
 import type { Context } from "../../context.js";
 import { TOKENS } from "../../di/token.js";
 
@@ -10,16 +10,7 @@ export const getAuthorDetailSchema = z.object({
 });
 export type GetAuthorDetailRequest = z.infer<typeof getAuthorDetailSchema>;
 
-export function getAuthorDetailValidator(args: unknown, ctx: Context) {
-	const logger = ctx.container.get(TOKENS.LOGGER);
-	const valid = getAuthorDetailSchema.safeParse(args);
-	if (!valid.success) {
-		const error = new ValidError(valid.error);
-		logger.debug("invalid request", { error });
-		throw error;
-	}
-	return valid.data;
-}
+export const getAuthorDetailValidator = zodValidator(getAuthorDetailSchema);
 
 export interface AuthorDetailResponse {
 	id: string;
