+++ src/main/apis/illusts/illust.delete.api.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { ValidError } from "../../../../features/shared/error/valid/index.js";
+import { zodValidator } from "../../../../features/shared/validation/index.js";
 import type { Context } from "../../context.js";
 import { TOKENS } from "../../di/token.js";
 
@@ -8,16 +8,7 @@ export const deleteIllustSchema = z.object({
 });
 export type DeleteIllustRequest = z.infer<typeof deleteIllustSchema>;
 
-export function deleteIllustValidator(args: unknown, ctx: Context) {
-	const logger = ctx.container.get(TOKENS.LOGGER);
-	const valid = deleteIllustSchema.safeParse(args);
-	if (!valid.success) {
-		const error = new ValidError(valid.error);
-		logger.debug("invalid request", { error });
-		throw error;
-	}
-	return valid.data;
-}
+export const deleteIllustValidator = zodValidator(deleteIllustSchema);
 
 export interface DeleteIllustResponse {
 	success: boolean;
