+++ src/main/apis/illusts/illust.register.api.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { createScopedContainer } from "../../../../features/shared/container/index.js";
-import { ValidError } from "../../../../features/shared/error/valid/index.js";
+import { zodValidator } from "../../../../features/shared/validation/index.js";
 import { Tag } from "../../../../features/tag/tag.model.js";
 import type { Context } from "../../context.js";
 import { TOKENS } from "../../di/token.js";
@@ -12,16 +12,7 @@ export const registerIllustSchema = z.object({
 });
 export type RegisterIllustRequest = z.infer<typeof registerIllustSchema>;
 
-export function registerIllustValidator(args: unknown, ctx: Context) {
-	const logger = ctx.container.get(TOKENS.LOGGER);
-	const valid = registerIllustSchema.safeParse(args);
-	if (!valid.success) {
-		const error = new ValidError(valid.error);
-		logger.debug("invalid request", { error });
-		throw error;
-	}
-	return valid.data;
-}
+export const registerIllustValidator = zodValidator(registerIllustSchema);
 
 export async function registerIllust(
 	ctx: Context,
