+++ src/main/apis/illusts/illust.search.api.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { ValidError } from "../../../../features/shared/error/valid/index.js";
+import { zodValidator } from "../../../../features/shared/validation/index.js";
 import { convertIllustArrayContentPathsToUrls } from "../../../server/helpers/illust.helper.js";
 import type { Context } from "../../context.js";
 import { TOKENS } from "../../di/token.js";
@@ -13,16 +13,7 @@ export const searchIllustSchema = z.object({
 });
 export type SearchIllustRequest = z.infer<typeof searchIllustSchema>;
 
-export function searchIllustValidator(args: unknown, ctx: Context) {
-	const logger = ctx.container.get(TOKENS.LOGGER);
-	const valid = searchIllustSchema.safeParse(args);
-	if (!valid.success) {
-		const error = new ValidError(valid.error);
-		logger.debug("invalid request", { error });
-		throw error;
-	}
-	return valid.data;
-}
+export const searchIllustValidator = zodValidator(searchIllustSchema);
 
 export async function searchIllust(ctx: Context, request: SearchIllustRequest) {
 	const { container } = ctx;
